requirements.txt
Django==3.2.13
django-ckeditor==6.3.2
django-allauth==0.51.0
djangocms-admin-style==2.0.2
pillow==8.4.0
psycopg2==2.9.3